pyproject.toml
[tool.poetry]
name = "cdp-agentkit-sample"
version = "0.1.0"
description = "A sample app that uses CDP AgentKit"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
cdp-langchain = "^0.0.6"
cdp-agentkit-core = "^0.0.4"
langchain-core = "^0.3.15"
langchain-openai = "^0.2.6"
langgraph = "^0.2.45"
pydantic = "^2.9.2"
cdp-sdk = "^0.10.0"
python-dotenv = "^1.0.1"
langchain-community = "^0.3.11"
requests = "^2.32.3"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
apscheduler = "^3.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
